Consider moving complex methods out of ArrayHelper #29
